The goal of this is to change nothing except the language this function is written in. That will make it easier to evolve in the future, plus reduces the overhead of `@threads`. Using the benchmark from #32181 (`func_threaded`):
Before, 1 thread: 17.740 μs
Before, 2 threads: 19.890 μs
After, 1 thread: 14.639 μs
After, 2 threads: 16.095 μs